Add koa middleware usage example
